+++ packages/components/atoms/f-link/src/tenants/en-GB.js
@@ -1,6 +1,7 @@
 export default {
     ariaDescription: {
         newLocation: 'Opens in a new window/screen/tab',
-        external: 'Opens an external site in a new window/screen/tab'
+        external: 'Opens an external site',
+        externalnewLocation: 'Opens an external site in a new window/screen/tab'
     }
 };