+++ packages/components/atoms/f-link/src/tenants/es-ES.js
@@ -1,6 +1,7 @@
 export default {
     ariaDescription: {
         newLocation: 'Abrir una nueva ventana/pestaña',
-        external: 'Abrir a una web externa en una nueva ventana/pestaña'
+        external: 'Abrir a una web externa',
+        externalnewLocation: 'Abrir a una web externa en una nueva ventana/pestaña'
     }
 };