+++ packages/components/atoms/f-link/src/tenants/it-IT.js
@@ -1,6 +1,7 @@
 export default {
     ariaDescription: {
         newLocation: 'Stai aprendo una nuova finestra/schermata/scheda',
-        external: 'Stai aprendo un sito esterno in una nuova finestra/schermata/scheda'
+        external: 'Stai aprendo un sito esterno',
+        externalnewLocation: 'Stai aprendo un sito esterno in una nuova finestra/schermata/scheda'
     }
 };