+++ packages/components/atoms/f-button/test-utils/component-objects/f-button.component.js
@@ -5,10 +5,20 @@ module.exports = class Buttons extends Page {
 
     get linkComponent () { return $('[data-test-id="link-button-component"]'); }
 
-    open (buttonType = '') {
-        const url = buttonType === 'link' ? '&knob-Button%20Type=link&knob-href=link' : '';
+    /**
+     * @description
+     * Sets the data for the button component.
+     * 
+     * @param {Object} button
+     * @param {String} button.type
+     * @param {String} button.size
+     */
+    open (button) {
+        const type = `&knob-Button%20Type=${button.type}`;
+        const url = button.type === 'link' ? `&knob-href=link` : '';
+        const size = `&knob-Button%20Size=${button.size}`;
 
-        browser.url(`/iframe.html?id=components-atoms-f-button--button-component${url}`)
+        browser.url(`/iframe.html?id=components-atoms-f-button--button-component${type}${url}${size}`)
     }
 
     waitForActionComponent () {